Restructure the optimized direct_spin1_cplx file
